fuck safe delete

+++ .aliases.sh
@@ -35,23 +35,22 @@ if command -v nvim > /dev/null 2>&1; then
     alias vim=nvim
 fi
 
+_has() {
+    type -v $1>/dev/null 2>&1
+}
+
 # Safe delete
-# mac
-if [ -d ~/.Trash ]; then
-    del() {
-	mv "$@" ~/.Trash
-    }
-else # probs on ubuntu
+if  _has gio; then
     if [ ! -d ~/.trash ]; then
         mkdir ~/.trash
     fi
 
-    if
     del(){
     	gio trash "$@"
     }
+
+    alias rm='echo "rm disabled! Use del for safe delete"'
 fi
-alias rm='echo "rm disabled! Use del for safe delete"'
 
 # More ls aliases.
 alias ll='ls -alF'
